% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_commune.R
\name{get_commune}
\alias{get_commune}
\title{Pobieranie granic gmin z PRG}
\usage{
get_commune(api_key)
}
\arguments{
\item{api_key}{twój klucz do API}
}
\value{
obiekt sf zawierający atrybuty i geometrię pobranych gmin
}
\description{
Funkcja służąca do pobierania granic wszystkich gmin
UWAGA - JEST TO DUŻY ZESTAW DANYCH
}
\examples{
get_commune('jEarp1RvgA4SgU3lWv3_sQ')
}
