% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get_country}
\alias{get_country}
\title{Pobieranie granic Polski z PRG}
\usage{
get_country()
}
\value{
obiekt sf zawierający atrybuty i geometrię dla Polski
}
\description{
Funkcja służąca do pobrania granic Polski
}
\examples{
\dontrun{
get_country()
}
}
