% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_county.R
\name{get_county}
\alias{get_county}
\title{Pobieranie granic powiatów z PRG}
\usage{
get_county(filter)
}
\arguments{
\item{filter}{filtr danych np. powiaty dla danego województwa (kod teryt)}
}
\value{
obiekt sf zawierający atrybuty i geometrię pobranych granic powiatów
}
\description{
Funkcja służąca do pobierania granic wszystkich powiatów.
UWAGA - JEST TO DUŻY ZESTAW DANYCH
}
\examples{
\dontrun{
#Pobranie wszystkich powiatów
get_county()
#Pobranie powiatów dla województwa wielkopolskiego
get_county('30')
}
}
