% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key.R
\name{get_key}
\alias{get_key}
\title{Wygenerowanie klucza api}
\usage{
get_key(email)
}
\arguments{
\item{email}{twój adres e-mail, na który zostanie wysłany klucz}
}
\description{
Funkcja służąca do wygenerowania klucza api
}
\examples{
get_key('email@email.com')
}
