% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Pobieranie terytów wybranych jednostek}
\usage{
get_meta(api_key, unit)
}
\arguments{
\item{api_key}{twój klucz do API}

\item{unit}{jednostka PRG}
}
\value{
dataframe zawierający nazwę oraz teryt wybranych jednostek
}
\description{
Funkcja służąca do pobierania terytów i nazw wybranych jednostek PRG
}
\examples{
get_meta('jEarp1RvgA4SgU3lWv3_sQ', 'wojewodztwa')
}
