% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prg-package.R
\docType{package}
\name{prgdownloader-package}
\alias{prgdownloader}
\alias{prgdownloader-package}
\title{prgdownloader: Pobieranie danych PRG}
\description{
Pobieranie danych z PRG
}
\author{
\strong{Maintainer}: Jakub Skowroński \email{skowronski.jakub97@gmail.com}

}
\keyword{internal}
