% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_key.R
\name{save_key}
\alias{save_key}
\title{Zapisanie klucza api}
\usage{
save_key(api_key)
}
\arguments{
\item{api_key}{klucz dla api prg}
}
\description{
Funkcja służąca do zapisania klucza api
}
\examples{
\dontrun{
save_key('123asd')
}
}
