% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_commune.R
\name{get_commune}
\alias{get_commune}
\title{Pobieranie granic gmin z PRG}
\usage{
get_commune(filter)
}
\arguments{
\item{filter}{filtr danych np. gminy dla danego powiatu (kod teryt)}
}
\value{
obiekt sf zawierający atrybuty i geometrię pobranych gmin
}
\description{
Funkcja służąca do pobierania granic wszystkich gmin.
UWAGA - JEST TO DUŻY ZESTAW DANYCH
}
\examples{
\dontrun{
#Pobranie wszystkich gmin
get_commune()
#Pobranie gmin dla powiatu chodzieskiego
get_commune('3001')
#Pobranie gmin dla województwa wielkopolskiego
get_commune('30')
}
}
