% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get_country}
\alias{get_country}
\title{Pobieranie granic Polski z PRG}
\usage{
get_country(api_key)
}
\arguments{
\item{api_key}{twój klucz do API}
}
\value{
obiekt sf zawierający atrybuty i geometrię dla Polski
}
\description{
Funkcja służąca do pobrania granic Polski
}
\examples{
get_country('jEarp1RvgA4SgU3lWv3_sQ')
}
