% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_county.R
\name{get_county}
\alias{get_county}
\title{Pobieranie granic powiatów z PRG}
\usage{
get_county(api_key)
}
\arguments{
\item{api_key}{twój klucz do API}
}
\value{
obiekt sf zawierający atrybuty i geometrię pobranych granic powiatów
}
\description{
Funkcja służąca do pobierania granic wszystkich powiatów
UWAGA - JEST TO DUŻY ZESTAW DANYCH
}
\examples{
get_county('jEarp1RvgA4SgU3lWv3_sQ')
}
