% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Pobieranie terytów wybranych jednostek}
\usage{
get_meta(unit)
}
\arguments{
\item{unit}{jednostka PRG}
}
\value{
dataframe zawierający nazwę oraz teryt wybranych jednostek
}
\description{
Funkcja służąca do pobierania terytów i nazw wybranych jednostek PRG
}
\examples{
\dontrun{
get_meta('wojewodztwa')
}
}
