% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_province.R
\name{get_province}
\alias{get_province}
\title{Pobieranie granic województw z PRG}
\usage{
get_province(api_key)
}
\arguments{
\item{api_key}{twój klucz do API}
}
\value{
obiekt sf zawierający atrybuty i geometrię pobranych województw
}
\description{
Funkcja służąca do pobierania granic wszystkich województw
UWAGA - JEST TO DUŻY ZESTAW DANYCH
}
\examples{
get_province('jEarp1RvgA4SgU3lWv3_sQ')
}
