% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prg-package.R
\docType{package}
\name{prgdownloader-package}
\alias{prgdownloader}
\alias{prgdownloader-package}
\title{prgdownloader: Pobieranie danych PRG}
\description{
Pobieranie danych z PRG
}
\author{
\strong{Maintainer}: Jakub Skowroński \email{skowronski.jakub97@gmail.com}

Authors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
}

}
\keyword{internal}
